Add natural scale summary parameters for a lognormal distribution

Add natural scale summary parameters for a lognormal distribution

Usage

add_natural_scale_mean_sd(dt)

Arguments

dt

...

See also