epidist: Estimate epidemiological delay distributions for infectious diseases

Understanding and accurately estimating epidemiological delay distributions is important for public health policy. These estimates directly influence epidemic situational awareness, control strategies, and resource allocation. In this package, we provide methods to address the key challenges in estimating these distributions, including truncation, interval censoring, and dynamical biases. Despite their importance, these issues are frequently overlooked, often resulting in biased conclusions.

Author

Maintainer: Sang Woo Park swp2@princeton.edu (ORCID)

Authors:

Sam Abbott contact@samabbott.co.uk (ORCID)

Adam Howes adamthowes@gmail.com (ORCID)