Define the model-specific component of an epidist custom formula for the marginal model

Define the model-specific component of an epidist custom formula for the marginal model

Usage

# S3 method for class 'epidist_marginal_model'
epidist_formula_model(data, formula, ...)

Arguments

data

An object with class corresponding to an implemented model.

formula

An object of class stats::formula or brms::brmsformula (or one that can be coerced to those classes). A symbolic description of the model to be fitted. A formula must be provided for the distributional parameter mu, and may optionally be provided for other distributional parameters.

...

Additional arguments passed to method.

See also