Transform data for the naive model

This method transforms data into the format required by the naive model by:

Identifying required columns for the naive model

Summarising the data by counting unique combinations of these columns and any variables in the model formula using .summarise_n_by_formula()

Converting the summarised data to a naive model object using new_epidist_naive_model()

Informing the user about any data aggregation that occurred using .inform_data_summarised()

Usage

# S3 method for class 'epidist_naive_model'
epidist_transform_data_model(data, family, formula, ...)

Arguments

data

The data to transform

family

The epidist family object specifying the distribution

formula

The model formula

...

Additional arguments passed to methods

See also