Estimate delays adjusted for right truncation and censoring using a latent model

Estimate delays adjusted for right truncation and censoring using a latent model

Usage

latent_truncation_censoring_adjusted_delay(
  formula = brms::bf(delay_central | vreal(obs_t, pwindow_upr, swindow_upr) ~ 1, sigma ~
    1),
  data,
  fn = brms::brm,
  family = brms::custom_family("latent_lognormal", dpars = c("mu", "sigma"), links =
    c("identity", "log"), lb = c(NA, 0), ub = c(NA, NA), type = "real", vars =
    c("pwindow", "swindow", "vreal1"), loop = FALSE),
  scode_functions =
    "\n    real latent_lognormal_lpdf(vector y, vector mu, vector sigma,\n                               vector pwindow, vector swindow,\n                               array[] real obs_t) {\n      int n = num_elements(y);\n      vector[n] d = y - pwindow + swindow;\n      vector[n] obs_time = to_vector(obs_t) - pwindow;\n      return lognormal_lpdf(d | mu, sigma) - \n        lognormal_lcdf(obs_time | mu, sigma);\n      }\n  ",
  scode_parameters =
    "\n    vector<lower = 0, upper = 1>[N] swindow_raw;\n    vector<lower = 0, upper = 1>[N] pwindow_raw;\n  ",
  scode_tparameters =
    "\n    vector<lower = 0>[N] pwindow;\n    vector<lower = 0>[N] swindow;\n    swindow = to_vector(vreal3) .* swindow_raw;\n    pwindow[noverlap] = to_vector(vreal2[noverlap]) .* pwindow_raw[noverlap];\n    if (wN) {\n      pwindow[woverlap] = swindow[woverlap] .* pwindow_raw[woverlap];\n    }\n  ",
  scode_priors =
    "\n    swindow_raw ~ uniform(0, 1);\n    pwindow_raw ~ uniform(0, 1);\n  ",
  ...
)