epidist: Estimate Epidemiological Delay Distributions With brms

Understanding and accurately estimating epidemiological delay distributions is important for public health policy. These estimates influence epidemic situational awareness, control strategies, and resource allocation. This package provides methods to address the key challenges in estimating these distributions, including truncation, interval censoring, and dynamical biases. These issues are frequently overlooked, resulting in biased conclusions. Built on top of 'brms', it allows for flexible modelling including time-varying spatial components and partially pooled estimates of demographic characteristics.

See also

Author

Maintainer: Sam Abbott contact@samabbott.co.uk (ORCID)

Authors:

Adam Howes adamthowes@gmail.com (ORCID)

Sang Woo Park swp2@princeton.edu (ORCID)

Other contributors:

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID) [contributor]